Objects now upload to the server. Added class for validating objects before they are uploaded to the server.

+++ src/iOS/Save-a-Selfie/Save a Selfie 2/GUI/SASMapViewController/SASFilterView/SASFilterView.m
@@ -75,7 +75,7 @@ SASDeviceType availableDevicesToFilter[5] = {
 
 #pragma mark UITableViewDelegate
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    return sizeof(availableDevicesToFilter) / sizeof(NSInteger);
+    return sizeof(availableDevicesToFilter) / sizeof(int);
 }
